refactor(grunt): move grunt tasks into separate module Fixes #17
