+++ grunttasks/blanket.js
@@ -1,0 +1,8 @@
+module.exports = function (grunt) {
+    grunt.config('blanket', {
+        'test-setup': {
+            src: ['lib/'],
+            dest: 'build/lib/'
+        }
+    });
+};
