+++ grunttasks/clean.js
@@ -1,0 +1,7 @@
+module.exports = function (grunt) {
+    grunt.config('clean', {
+        build: {
+            src: ['build/**']
+        }
+    });
+};
