+++ grunttasks/jshint.js
@@ -1,0 +1,10 @@
+module.exports = function (grunt) {
+    grunt.config('jshint', {
+        sdk: {
+            src: ['*.js', 'lib/**/*.js']
+        },
+        options: {
+            jshintrc: true
+        }
+    });
+};
