+++ package.json
@@ -3,57 +3,62 @@
   "description": "AWS Mobile Analytics SDK for JavaScript",
   "version": "0.9.2",
   "author": {
-    "name":"Amazon Web Services",
-    "email":"",
-    "url":"http://aws.amazon.com/"
+    "name": "Amazon Web Services",
+    "email": "",
+    "url": "http://aws.amazon.com/"
   },
   "homepage": "https://github.com/aws/aws-sdk-mobile-analytics-js",
-  "contributors": [
-  ],
+  "contributors": [],
   "devDependencies": {
-    "grunt": "~0.4.5",
-    "grunt-contrib-jshint": "~0.11.0",
-    "grunt-contrib-clean": "~0.6.0",
-    "grunt-contrib-copy": "~0.8.0",
-    "grunt-contrib-uglify": "~0.8.0",
-    "grunt-blanket": "~0.0.8",
-    "grunt-mocha-test": "~0.12.7",
-    "grunt-mocha": "~0.4.12",
-    "grunt-browserify": "~3.5.0",
-    "mocha-cobertura-reporter": "~1.0.4",
-    "travis-cov": "~0.2.5",
+    "chai": "*",
     "coffee-script": "~1.9.1",
     "coffeeify": "1.0.0",
+    "grunt": "~0.4.5",
+    "grunt-blanket": "~0.0.8",
+    "grunt-browserify": "~3.5.0",
+    "grunt-contrib-clean": "~0.6.0",
+    "grunt-contrib-copy": "~0.8.0",
+    "grunt-contrib-jshint": "~0.11.0",
+    "grunt-contrib-uglify": "~0.8.0",
+    "grunt-mocha": "~0.4.12",
+    "grunt-mocha-test": "~0.12.7",
+    "load-grunt-tasks": "3.5.0",
+    "mocha-cobertura-reporter": "~1.0.4",
     "semver": "*",
-    "chai": "*",
+    "travis-cov": "~0.2.5",
     "xml2js": "0.2.6"
   },
   "dependencies": {
     "aws-sdk": ">=2.2.37"
   },
   "main": "lib/ama.js",
-  "directories":  {
+  "directories": {
     "lib": "lib"
   },
   "engines": {
     "node": ">= 0.8.0"
   },
   "repository": {
-    "type" : "git",
+    "type": "git",
     "url": "git://github.com/aws/aws-sdk-mobile-analytics-js"
   },
-  "bugs" : {
-    "url" : "http://github.com/aws/aws-sdk-mobile-analytics-js/issues",
-    "mail" : ""
+  "bugs": {
+    "url": "http://github.com/aws/aws-sdk-mobile-analytics-js/issues",
+    "mail": ""
   },
-  "licenses": [{
-    "type": "Apache 2.0",
-    "url": "http://github.com/aws/aws-sdk-mobile-analytics-js/raw/master/LICENSE.txt"
-  }],
-  "keywords": [
-    "api", "amazon","aws","mobileanalytics"
+  "licenses": [
+    {
+      "type": "Apache 2.0",
+      "url": "http://github.com/aws/aws-sdk-mobile-analytics-js/raw/master/LICENSE.txt"
+    }
   ],
-  "scripts" : {
+  "keywords": [
+    "api",
+    "amazon",
+    "aws",
+    "mobileanalytics"
+  ],
+  "scripts": {
     "test": "grunt test",
     "unit": "grunt test",
     "coverage": "grunt test-nodejs",
